tasks redirect

+++ app/Http/Controllers/TasksController.php
@@ -55,10 +55,15 @@ class TasksController extends Controller
     public function show($id)
     {
         $task = Task::find($id);
-
+        
+        if (\Auth::id() === $task->user_id) {
         return view('tasks.show', [
             'task' => $task,
         ]);
+        }else{
+            return redirect('/');
+        }
+        
     }
 
     // getでtasks/id/editにアクセスされた場合の「更新画面表示処理」
@@ -66,9 +71,14 @@ class TasksController extends Controller
     {
         $task = Task::find($id);
 
-        return view('tasks.edit', [
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', [
             'task' => $task,
         ]);
+        }else{
+            return redirect('/');
+        }
+        
     }
 
     // putまたはpatchでtasks/idにアクセスされた場合の「更新処理」
@@ -79,10 +89,15 @@ class TasksController extends Controller
             'content' => 'required|max:191',
         ]);
         
-        $task = Task::find($id);
+         if (\Auth::id() === $task->user_id) {
+            $task = Task::find($id);
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
+        }else{
+            return redirect('/');
+        }
+        
 
         return redirect('/');
     }
@@ -90,9 +105,12 @@ class TasksController extends Controller
     // deleteでtasks/idにアクセスされた場合の「削除処理」
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
+        $task = \App\Task::find($id);
 
-        return redirect('/');
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+
+        return back();
     }
 }
